backend/django_admin/requirements.txt
Django>=4.2,<5.0
psycopg2-binary>=2.9.9
Pillow>=10.0.0
python-dotenv>=1.0.0
django-cors-headers>=4.3.0
django-import-export>=3.3.0
django-filter>=23.5
djangorestframework>=3.14.0
